Final
it's a final version of this Lab i think. I added the load to file and from it. Also added something like a menu and small fixes to existing functions. For example, now if you remove an edge from the matrix, the message will be different depending on whether it existed before or not. Some gramma correction etc.